requirements.txt
absl-py==0.11.0
asgiref==3.3.1
astunparse==1.6.3
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
flatbuffers==1.12
gast==0.3.3
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==2.10
image==1.5.33
importlib-metadata==3.4.0
Keras-Preprocessing==1.1.2
Markdown==3.3.3
MouseInfo==0.1.3
numpy==1.19.4
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
pywin32 @ file:///C:/Users/Denny/Desktop/CASPER/pywin32-300-cp37-cp37m-win_amd64.whl
Pillow==8.0.1
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyQt5==5.15.2
PyQt5-sip==12.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7
six==1.15.0
sqlparse==0.4.1
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.7.0
termcolor==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
Werkzeug==1.0.1
wrapt==1.12.1
wxPython==4.1.1
xxhash==2.0.0
zipp==3.4.0
textdetection~=0.2.1
torch~=1.8.1
pandas~=1.1.5
transformers~=4.6.1
configparser~=5.0.2
pytesseract~=0.3.7
sklearn~=0.0
scikit-learn~=0.24.2